pyproject.toml
[project]
name = "watch-recognition"
version = "0.1.0"
description = "This project aims to read time from analog watches."
readme = "README.md"
requires-python = ">=3.12.7"
dependencies = [
    "dvc>=3.62.0",
    "ipykernel>=6.30.1",
    "label-studio>=1.20.0",
    "matplotlib>=3.10.5",
    "pyyaml>=6.0.2",
    "scikit-image>=0.25.2",
    "scipy>=1.16.1",
]
